pyproject.toml
[tool.poetry]
name = "asamint"
version = "0.1.4"
description = "Integration package for ASAM related projects"
authors = ["Christoph Schueler <cpu12.gems@googlemail.com>"]
license = "GPLv2"
readme = "docs/README.rst"
keywords = ["automotive", "ecu", "xcp", "asam", "autosar"]
homepage = "https://github.com/christoph2/asamint"
repository = "https://github.com/christoph2/asamint"
# documentation = "https://github.com/christoph2/asamint/wiki"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers", "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
	"Programming Language :: Python :: 3.13"
]
# build = "build_ext.py"
include = [
    { path = "asamint/data/*.*", format = "wheel" },

 #   { path = "CMakeLists.txt", format = "sdist" },

    { path = "asamint/data/*.*", format = "sdist" },
]


[tool.poetry.dependencies]
python = "^3.9"
pya2ldb = "^0.13.1"
objutils = "^0.4.12"
asammdf = "^8.0.1"
rich = "^13.9.2"
traitlets = "^5.11.2"
pyxcp = "^0.22.9"
sortedcontainers = "^2.4.0"
lxml = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
black = "^24.10.0"
pre-commit = "^4.0.1"
pre-commit-hooks = "^5.0.0"
pybind11 = "^2.13.6"

[build-system]
requires = ["poetry-core>=1.0.0", "pybind11>=2.12.0",  "pybind11[global]>=2.12.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "asamint"
dynamic = ["license", "readme", "authors", "requires-python", "description", "classifiers", "scripts", "dependencies", "optional-dependencies"]

[tool.pytest]
addopts = "--verbose --tb=short --junitxml=result.xml -o junit_family=xunit2"
testpaths = "pyxcp/tests"

[tool.isort]
profile = "black"
force_single_line = false
lines_after_imports = 2

[tool.mypy]
strict = false
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true

[tool.flake8]
ignore = ["D203", "E203", "E266", "E501", "W503", "F403", "F401", "BLK100"]
exclude = [
    ".git",
    "__pycache__",
    " __pypackages__",
    ".mypy_cache",
    ".tox",
    ".venv",
    ".eggs",
    "_build",
    "build",
    "docs",
    "dist",
    "experimental",
	"asamint/parsers/*",
    "asamint/damos/parsers/*"
]

max-complexity = 10
count = true
statistics = true
show-source = true
max-line-length = 132
select = ["B","C","E","F","W","T4","B9"]
# extend-select = "B950"
extend-ignore = ["E203", "E501", "E701"]

[tool.ruff]
line-length = 132

[tool.black]
line-length=132
include = '\.pyi?$'
exclude = [
    ".git",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "build",
    "docs",
    "experimental",
    "__pycache__",
    "__pypackages__",
    "dist",
	"asamint/parsers/*",
    "asamint/damos/parsers/*",
]
